refactor: rename erc4337 ox type shim to inference
